requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.5b0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.2
django-channels==0.7.0
django-redis==4.12.1
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
zope.interface==5.4.0